+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add TUI switch for status and priority
 
+### Changed
+
+- Set default status for todo if not available
+
 ## [v0.2.1] - 2025-07-27
 
 ### Added
@@ -28,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- Removed intuitive priority alias
+- Remove intuitive priority alias
 
 ## [v0.2.0] - 2025-07-25